pyproject.toml
[project]
name = "stock-pipeline"
version = "0.1.0"
description = ""
authors = [
    {name = "Poornachandra.A.N",email = "poornachandra308@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "apache-airflow-providers-postgres (>=6.2.3,<7.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
